Convert replacePaths script to an ES module
