+++ package.json
@@ -48,7 +48,7 @@
   },
   "scripts": {
     "clean": "rimraf coverage dist tmp",
-    "build": "tsc -p tsconfig.release.json && node lib/replacePaths.js",
+    "build": "tsc -p tsconfig.release.json && node lib/replacePaths.mjs",
     "build:watch": "tsc -w -p tsconfig.release.json",
     "format:check": "prettier --check \"**/**.ts\"",
     "format": "prettier --write \"**/**.ts\"",
